move mappings from SimStatus into new file to fix tests
