+++ js/menu/filter/filter.js
@@ -5,6 +5,7 @@ import { reconnect } from "./reconnect.js";
 import { getVisible } from "../../events.js";
 import { units } from "../../types/units.js";
 import { copyObject } from "../../lib/copy.js";
+import { SimStatusBitFieldDisplayValues } from "../../../mappings/sim-status.js";
 
 const filterButton = document.getElementById("filter-button");
 const openFilter = document.getElementById("open-filter");
@@ -59,17 +60,6 @@ let parametersRange = units.sort((a, b) =>
 );
 
 parametersRange = parametersRange.map((parameter) => new Range(parameter));
-
-export const SimStatusBitFieldDisplayValues = {
-  23: "Overlay",
-  24: "Stopped",
-  25: "LeftDetector",
-  26: "DecayedInCalorimeter",
-  27: "DecayedInTracker",
-  28: "VertexIsNotEndpointOfParent",
-  29: "Backscatter",
-  30: "CreatedInSimulation",
-};
 
 const bits = new BitFieldBuilder(
   "simulatorStatus",
