+++ mappings/sim-status.js
@@ -1,0 +1,10 @@
+export const SimStatusBitFieldDisplayValues = {
+  23: "Overlay",
+  24: "Stopped",
+  25: "LeftDetector",
+  26: "DecayedInCalorimeter",
+  27: "DecayedInTracker",
+  28: "VertexIsNotEndpointOfParent",
+  29: "Backscatter",
+  30: "CreatedInSimulation",
+};
